feat: extracted db migrations from the coordinator and transaction-exclusion app
