+++ coordinator/ethereum/blob-submitter/build.gradle
@@ -24,6 +24,7 @@ dependencies {
 
   testImplementation(project(":jvm-libs:linea:testing:l1-blob-and-proof-submission"))
   testImplementation(project(":coordinator:persistence:aggregation"))
+  testImplementation(project(":coordinator:persistence:db-common"))
   testImplementation(testFixtures(project(":jvm-libs:persistence:db")))
   testImplementation(testFixtures(project(":coordinator:ethereum:gas-pricing")))
   testImplementation(testFixtures(project(":coordinator:core")))
