+++ coordinator/persistence/aggregation/build.gradle
@@ -10,6 +10,7 @@ dependencies {
 
   testImplementation(project(":coordinator:persistence:batch"))
   testImplementation(project(":coordinator:persistence:blob"))
+  testImplementation(project(":coordinator:persistence:db-common"))
   testImplementation(testFixtures(project(":jvm-libs:persistence:db")))
   testImplementation(testFixtures(project(":coordinator:core")))
   testImplementation(testFixtures(project(":jvm-libs:generic:extensions:kotlin")))
