+++ coordinator/persistence/batch/src/integrationTest/kotlin/net/consensys/zkevm/persistence/dao/batch/persistence/BatchesPostgresDaoTest.kt
@@ -28,7 +28,6 @@ import kotlin.time.Duration.Companion.seconds
 class BatchesPostgresDaoTest : CleanDbTestSuiteParallel() {
   init {
     target = "4"
-    migrationLocations = "filesystem:../../app/src/main/resources/db/"
   }
 
   override val databaseName = DbHelper.generateUniqueDbName("coordinator-tests-batches")
