+++ coordinator/persistence/db-common/build.gradle
@@ -1,0 +1,3 @@
+plugins {
+  id "net.consensys.zkevm.kotlin-library-conventions"
+}
