+++ coordinator/persistence/db-common/src/main/resources/db/V002__add_blob_and_aggregation_schema.sql
