+++ coordinator/persistence/db-common/src/main/resources/db/V003__add_feehistory_schema.sql
