+++ coordinator/persistence/feehistory/build.gradle
@@ -12,6 +12,7 @@ dependencies {
   implementation(project(":jvm-libs:core:domain-models"))
   implementation(project(":coordinator:ethereum:gas-pricing:dynamic-cap"))
 
+  testImplementation(project(":coordinator:persistence:db-common"))
   testImplementation("com.fasterxml.jackson.core:jackson-databind:${libs.versions.jackson.get()}")
   testImplementation("com.fasterxml.jackson.core:jackson-annotations:${libs.versions.jackson.get()}")
   testImplementation("com.fasterxml.jackson.module:jackson-module-kotlin:${libs.versions.jackson.get()}")
