+++ jvm-libs/persistence/db/build.gradle
@@ -19,7 +19,9 @@ dependencies {
     because("Flyway DB and other dependencies use SLF4J")
   }
 
+  testImplementation(project(":coordinator:persistence:db-common"))
   testImplementation testFixtures(project(':jvm-libs:generic:extensions:kotlin'))
+
   testFixturesImplementation(project(":coordinator:core"))
   testFixturesApi(platform("org.junit:junit-bom:${libs.versions.junit.get()}"))
   testFixturesApi("io.vertx:vertx-junit5:${libs.versions.vertx.get()}")
