+++ settings.gradle
@@ -46,6 +46,7 @@ include 'coordinator:persistence:blob'
 include 'coordinator:persistence:aggregation'
 include 'coordinator:persistence:batch'
 include 'coordinator:persistence:feehistory'
+include 'coordinator:persistence:db-common'
 include 'coordinator:ethereum:gas-pricing'
 include 'coordinator:ethereum:gas-pricing:static-cap'
 include 'coordinator:ethereum:gas-pricing:dynamic-cap'
