+++ transaction-exclusion-api/persistence/rejectedtransaction/src/main/resources/db/V001__initial_schema.sql
